requirements.txt
pandas==1.5
pyomo==6.5
openpyxl==3.1
matplotlib==3.7
PyYAML==6.0
dash==2.7.0
dash-bootstrap-components==1.2.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
numpy==1.21.6
plotly==5.7.0
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.15
scipy==1.7.3
gunicorn
